Separacion de actividad con clases y adaptadores
