+++ app/src/main/java/com/example/vitalage/model/MedicationCard.kt
@@ -1,0 +1,10 @@
+package com.example.vitalage.model
+
+data class MedicationCard(
+    val name: String,
+    val dose: String,
+    val observation: String,
+    val morning: String,
+    val afternoon: String,
+    val night: String
+)
